implement fetchBlocks for optimize unprocessedBlocks
